Extracted Actions to common folder for additional usage in UserRepositories
