+++ src/common/Actions/Idle/index.js
@@ -1,6 +1,6 @@
 import { MapIcon } from "../styled";
-import { CommonContainer } from "../../../../../common/Containers/styled";
-import { CommonParagraph } from "../../../../../common/Texts/styled";
+import { CommonContainer } from "../../Containers/styled";
+import { CommonParagraph } from "../../Texts/styled";
 
 const Idle = () => (
     <CommonContainer $action>
